Add COOKIE_SECRET support and fix incorrect Redis/Mongo URIs
